+++ src/ifopt_core/include/ifopt/composite.h
