+++ src/ifopt_core/include/ifopt/cost_term.h
