+++ src/ifopt_core/include/ifopt/variable_set.h
