+++ src/ifopt_core/src/composite.cc
