+++ src/ifopt_core/src/leaves.cc
