+++ src/ifopt_core/test/composite_test.cc
