+++ src/ifopt_ipopt/include/ifopt/ipopt_adapter.h
