+++ src/ifopt_ipopt/src/ipopt_adapter.cc
@@ -24,7 +24,7 @@ OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************************************************************/
 
-#include <ifopt_ipopt/ipopt_adapter.h>
+#include <ifopt/ipopt_adapter.h>
 
 namespace ifopt {
