+++ src/ifopt_snopt/include/ifopt/snopt_adapter.h
