+++ src/ifopt_snopt/src/snopt_adapter.cc
@@ -24,7 +24,7 @@ OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************************************************************/
 
-#include <ifopt_snopt/snopt_adapter.h>
+#include <ifopt/snopt_adapter.h>
 
 namespace ifopt {
